+++ changelog.md
@@ -5,6 +5,7 @@
 - Search on map now zooms to and animates the found item
 - Swipe to remove favorite camps, events, art
 - Double tap on map for zoom in/far/out
+- Fuzzy search for art, events and camps
 
 ## 2.37
 - Search for camps, events and art is sorted by importance